requirements.txt
fastapi==0.60.1
pydantic==1.5.1
asyncpg==0.20.1
uvicorn==0.11.5
SQLAlchemy==1.3.17
psycopg2-binary==2.8.5
gino==1.0.0
alembic==1.4.2
pytest==5.4.3
passlib==1.7.2
python-jose==3.1.0
